refactor: 상품 도메인과 엔티티 분리
