+++ src/main/java/cart/domain/Product.java
@@ -2,27 +2,14 @@ package cart.domain;
 
 public class Product {
 
-    private final Long id;
     private final String name;
     private final int price;
     private final String imageUrl;
 
-    public Product(final Long id, final String name, final int price, final String imageUrl) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
     public Product(final String name, final int price, final String imageUrl) {
-        this.id = null;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
-    }
-
-    public Long getId() {
-        return id;
     }
 
     public String getName() {
